% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CompleteI.R
\name{CompleteI}
\alias{CompleteI}
\title{complete intensity information
the extract intensity may be null or more than one. Need to give mz (mz.lib) and int (0) to NULL. Selet the higher intensity if more than 1}
\usage{
CompleteI(record, mz.lib)
}
\arguments{
\item{record}{a matrix of m/z and intensity}

\item{the}{mz in library}
}
\description{
complete intensity information
the extract intensity may be null or more than one. Need to give mz (mz.lib) and int (0) to NULL. Selet the higher intensity if more than 1
}

