% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CompleteMs1Scan.R
\name{CompleteMs1Scan}
\alias{CompleteMs1Scan}
\title{Complete MS1 scans
MS1 scans could be lost, find missing MS1 scan and complete with 0}
\usage{
CompleteMs1Scan(scantime.ms1.raw, scan.ms1.raw)
}
\arguments{
\item{scantime.ms1.raw}{raw MS1 scan time}

\item{scan.ms1.raw}{raw MS1 scans, could contain missing scans}
}
\description{
Complete MS1 scans
MS1 scans could be lost, find missing MS1 scan and complete with 0
}

