% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CompleteMs1Scantime.R
\name{CompleteMs1Scantime}
\alias{CompleteMs1Scantime}
\title{Complete MS1 scantime
MS1 scantime could be lost, find missing MS1 scantime and complete with 0}
\usage{
CompleteMs1Scantime(scantime.ms1.raw)
}
\arguments{
\item{scantime.ms1.raw}{MS1 scan time, could contain missing scantime}
}
\description{
Complete MS1 scantime
MS1 scantime could be lost, find missing MS1 scantime and complete with 0
}

