% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CompleteMs2Scan.R
\name{CompleteMs2Scan}
\alias{CompleteMs2Scan}
\title{Complete MS2 scans
MS2 scans could be lost, find missing MS2 scan and complete with 0}
\usage{
CompleteMs2Scan(scantime.ms1, scantime.ms2, scan.ms2)
}
\arguments{
\item{scantime.ms1}{MS1 scan time}

\item{scantime.ms2}{MS2 scan time}

\item{scan.ms2}{MS2 scans}
}
\description{
Complete MS2 scans
MS2 scans could be lost, find missing MS2 scan and complete with 0
}

