% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CorPreFrag.R
\name{CorPreFrag}
\alias{CorPreFrag}
\title{Correlation between precursor and fragment ion
 Calculate correlation between precursor and fragment ions with Pearsom correlation}
\usage{
CorPreFrag(premzi, fragmzi)
}
\arguments{
\item{premzi}{a list contains a dataframe of precursor ion m/z and intensity}

\item{fragmzi}{a list contains a dataframe of precursor ion m/z and intensity}
}
\description{
Correlation between precursor and fragment ion
 Calculate correlation between precursor and fragment ions with Pearsom correlation
}

