% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ExtractMzI.R
\name{ExtractMzI}
\alias{ExtractMzI}
\title{Extract extract m/z and intensity from scan
applied on the extraction of both precursor ions and fragment ions}
\usage{
ExtractMzI(scan.all, scan.index, mz.lib, ppm)
}
\arguments{
\item{scan.all}{all scans of ms1 or ms2}

\item{scan.index}{scan index selected previously}

\item{mz.lib}{mz of targeted ions in library}

\item{ppm}{the given tolerance}
}
\description{
Extract extract m/z and intensity from scan
applied on the extraction of both precursor ions and fragment ions
}

