% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FindPeaksIdx.R
\name{FindPeaksIdx}
\alias{FindPeaksIdx}
\title{find peak index of matched precursor ions with given ppm
the precursor m/z below 400 is considered as 400 due to heavy mass shift in low mass range}
\usage{
FindPeaksIdx(peaktable, premz.lib, ms1ppm)
}
\arguments{
\item{peaktable}{provide the precursor ion m/z}

\item{premz.lib}{the precursor ion m/z in library}

\item{ms1ppm}{the tolerance precursor ion match}
}
\description{
find peak index of matched precursor ions with given ppm
the precursor m/z below 400 is considered as 400 due to heavy mass shift in low mass range
}

