% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FragMzI.R
\name{FragMzI}
\alias{FragMzI}
\title{Extract m/z and intensity of fragment ion
 Select SWATH window automatically according to precursor ion m/z, the extract fragment ion information}
\usage{
FragMzI(scan.ms2, scan.idx, premz.lib, fragmz.lib, ms2ppm, scantime.ms1,
  massrange.ms1, windows = file.windows)
}
\arguments{
\item{scan.ms2}{all ms2 scan}

\item{premz.lib}{precursor ion mz in library}

\item{fragmz.lib}{fragment ion mz in library}

\item{ms2ppm}{m/z tolerance of fragment ions}

\item{scantime.ms1}{ms1 scantime}

\item{massrange.ms1}{ms1 mass range}

\item{scan.index}{scan index selected previously}
}
\description{
Extract m/z and intensity of fragment ion
 Select SWATH window automatically according to precursor ion m/z, the extract fragment ion information
}

