% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MetDIA.R
\name{MetDIA}
\alias{MetDIA}
\title{MetDIA: MetDIA: Targeted Metabolite Extraction of Multiplexed MS/MS Spectra Generated by Data-Independent Acquisition
MetDIA: MetDIA: Targeted Metabolite Extraction of Multiplexed MS/MS Spectra Generated by Data-Independent Acquisition}
\usage{
MetDIA(filepath, speclib, ppm.pd = 25, sn = 6, peakwidth = c(5, 30),
  ms1ppm = 25, ms2ppm = 35, cutoff = 0.8, inf.detail = metinf.detail,
  refine = TRUE, ms2.quant = FALSE)
}
\arguments{
\item{filepath}{file path of data. different group data should be put into different folder}

\item{ppm.pd}{tolerance for peak detection}

\item{sn}{singal to noise ratio for peak detection}

\item{peakwidth}{peak width for peak detection}

\item{ms1ppm}{tolerance for MS1 match}

\item{ms2ppm}{tolerance for MS2 match}

\item{cutoff}{minium MCI score cutoff}

\item{inf.detail}{detail inforamtion for refine result, containing metabolite name, lab ID, KEGG ID, HMDB ID and so on}

\item{refine}{whether refine the result or not}

\item{ms2.quant}{whether output ms2 quant result}

\item{spcelib}{library information contains precursor ion m/z, fragment ion of m/z and relative intensity, name}
}
\description{
MetDIA: MetDIA: Targeted Metabolite Extraction of Multiplexed MS/MS Spectra Generated by Data-Independent Acquisition
MetDIA: MetDIA: Targeted Metabolite Extraction of Multiplexed MS/MS Spectra Generated by Data-Independent Acquisition
}

