% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Ms2MaxI.R
\name{Ms2MaxI}
\alias{Ms2MaxI}
\title{Extract fragment ion intensity at summit
 Extract fragment ion intensity at the MS1 EIC summit}
\usage{
Ms2MaxI(premzi, fragmzi)
}
\arguments{
\item{premzi}{a list contains a dataframe of precursor ion m/z and intensity}

\item{fragmzi}{a list contains a dataframe of precursor ion m/z and intensity}
}
\description{
Extract fragment ion intensity at summit
 Extract fragment ion intensity at the MS1 EIC summit
}

