% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PeakGroupRaw.R
\name{PeakGroupRaw}
\alias{PeakGroupRaw}
\title{Generate the peakgroup with raw rentiontime
the retention time of peakgroup generated from xset3 (filled peaks) is aligned, need the raw retention time for scan extraction}
\usage{
PeakGroupRaw(xset.fillpeaks = xset3, groupindex.single = groupindex.single)
}
\arguments{
\item{xset.fillpeaks}{the xset with filled gaps}

\item{groupindex.single}{one sample only contributes one peak to a peak group}
}
\description{
Generate the peakgroup with raw rentiontime
the retention time of peakgroup generated from xset3 (filled peaks) is aligned, need the raw retention time for scan extraction
}

