% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PKtoDP.R
\name{PKtoDP}
\alias{PKtoDP}
\title{Extract peak information and calculate score
Use precursor ion m/z in library to match the detected peaks. Then targeted extract m/z and intensity of precursor and fragment ions. Calculate dot product and correlation with these data}
\usage{
PKtoDP(peaktable, peaktable.corrected, scantime, speclib.single, scan.ms1,
  scan.ms2, ms1ppm, ms2ppm, massrange.ms1, peakgroup.num, mcicutoff = cutoff,
  windows = file.windows)
}
\arguments{
\item{peaktable}{peaktable with raw retention time}

\item{peaktable.corrected}{peaktable with corrected retention time}

\item{scantime}{MS1 scan time}

\item{speclib.single}{metabolite information in library, contains PrecursorMZ, ProductMz, LibraryIntensity and Name}

\item{scan.ms1}{MS1 scan, contain m/z and intensity}

\item{scan.ms2}{MS2 scan, contain m/z and intensity}

\item{ms1ppm}{tolerance for precursor ion match}

\item{ms2ppm}{tolerance for precursor ion match}

\item{massrange.ms1}{MS1 mass range}

\item{peakgroup.num}{for plot}

\item{mcicutoff}{the MCI score cutoff}
}
\description{
Extract peak information and calculate score
Use precursor ion m/z in library to match the detected peaks. Then targeted extract m/z and intensity of precursor and fragment ions. Calculate dot product and correlation with these data
}

