% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PlotAllRaw.R
\name{PlotAllRaw}
\alias{PlotAllRaw}
\title{Mirror plot and peak group plot with raw intensity data for all peaks
 Use m/z ang intensity of precursor and fragment ions for mirror plot and peak group plot. The left is peak group plot with all ions, the black line stands for precursor ion, and the colorful lines stand for fragment ions. The right is the mirror plot of all fragmet ions.}
\usage{
PlotAllRaw(ms1, ms2, feature, corvalue, ms2.maxi, ms2.maxi.cutoff, dpvalue,
  peakgroup.num)
}
\arguments{
\item{ms1}{precursor ion m/z and raw intensity}

\item{ms2}{fragment ion m/z and raw intensity}

\item{feature}{spectrum information in library}

\item{corvalue}{correlation value}

\item{ms2.maxi}{MS2 intensity at summit}

\item{ms2.maxi.cutoff}{MS2 intensity filtered by correlation (0.6) at summit}

\item{dpvalue}{dot product value}

\item{peakgroup.num}{differentiate from different samples}
}
\description{
Mirror plot and peak group plot with raw intensity data for all peaks
 Use m/z ang intensity of precursor and fragment ions for mirror plot and peak group plot. The left is peak group plot with all ions, the black line stands for precursor ion, and the colorful lines stand for fragment ions. The right is the mirror plot of all fragmet ions.
}

