% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PreMzI.R
\name{PreMzI}
\alias{PreMzI}
\title{Extract m/z and intensity of precursor ion
 Extract m/z and intensity of precursor ion and complete missing intensity}
\usage{
PreMzI(scan.ms1, scan.index, premz.lib, ms1ppm)
}
\arguments{
\item{scan.ms1}{all ms1 scan}

\item{scan.index}{scan index selected previously}

\item{premz.lib}{precursor ion mz in library}

\item{ms1ppm}{m/z tolerance of precursor ions}
}
\description{
Extract m/z and intensity of precursor ion
 Extract m/z and intensity of precursor ion and complete missing intensity
}

