% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RefineReuslt.ms2.R
\name{RefineReuslt.ms2}
\alias{RefineReuslt.ms2}
\title{RefineReuslt.ms2: Refine output result of MS2 area
RefineReuslt.ms2: Add detailed information, such as KEGG, HMDB, to the result}
\usage{
RefineReuslt.ms2(result.raw, inf.detail, name.sample)
}
\arguments{
\item{result.raw}{raw result of MS2 area}

\item{inf.detail}{detail inforamtion containing metabolite name, lab ID, KEGG ID, HMDB ID}

\item{name.sample}{sample name}
}
\description{
RefineReuslt.ms2: Refine output result of MS2 area
RefineReuslt.ms2: Add detailed information, such as KEGG, HMDB, to the result
}

