% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RefineReuslt.R
\name{RefineReuslt}
\alias{RefineReuslt}
\title{RefineReuslt: Refine output result
RefineReuslt: Add detailed information, such as KEGG, HMDB, to the result}
\usage{
RefineReuslt(result.raw, inf.detail, name.sample)
}
\arguments{
\item{result.raw}{raw result of MS1 and SWATH ID}

\item{inf.detail}{detail inforamtion containing metabolite name, lab ID, KEGG ID, HMDB ID}

\item{name.sample}{sample name}
}
\description{
RefineReuslt: Refine output result
RefineReuslt: Add detailed information, such as KEGG, HMDB, to the result
}

