% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ScanIdx.R
\name{ScanIdx}
\alias{ScanIdx}
\title{extract scan index with rt of matched ms1 feature
extract scan index with rt of matched ms1 feature. If without any match, return NA}
\usage{
ScanIdx(scantime, ms1.mzrt)
}
\arguments{
\item{scantime}{all ms1 scantime}

\item{ms1.mzrt}{mz and rt information of ms1matched ms1 feature}
}
\description{
extract scan index with rt of matched ms1 feature
extract scan index with rt of matched ms1 feature. If without any match, return NA
}

